Switch camera pose from SE3 to quaternion, also support multiple camera poses for later dynamic object. (#61)
Switch camera pose from SE3 to quaternion, to simplify later pose
optimization.
Also, a simple idea is to assign an object id to each point, take some
6DOF pose estimation for moving objects as input, then handle rigid
objects as points with different camera poses (movement of the object is
transformed into the movement of the camera). In this PR, we first
change the input format for rasterization. Will try to support the whole
feature in later PR.